feat: change editor node to inherit code node
