+++ packages/milkup-equations/src/EquationsPlugin.tsx
@@ -4,16 +4,12 @@ import {
   $createParagraphNode,
   $getSelection,
   $insertNodes,
-  $isBlockElementNode,
-  $isParagraphNode,
   $isRangeSelection,
   $isRootOrShadowRoot,
   $isTextNode,
   COMMAND_PRIORITY_EDITOR,
-  COMMAND_PRIORITY_LOW,
   createCommand,
   ElementNode,
-  KEY_BACKSPACE_COMMAND,
   LexicalCommand,
   TextNode,
 } from "lexical";
@@ -21,13 +17,14 @@ import {
   $createBlockEquationNode,
   $isBlockEquationNode,
   BlockEquationNode,
-} from "./BlockEquationNode";
-import { $isEquationEditorNode, EquationEditorNode } from "./EquationEditorNode";
-import { BlockEquationRendererNode } from "./BlockEquationRendererNode";
+} from "./block/BlockEquationNode";
+import {
+  $isEquationEditorNode,
+  EquationEditorNode,
+} from "./block/EquationEditorNode";
+import { BlockEquationRendererNode } from "./block/BlockEquationRendererNode";
 import { useEffect, useRef } from "react";
-import { $getNodeByKey } from 'lexical';
-
-export let hideActiveEquationEditor = () => {};
+import { $getNodeByKey } from "lexical";
 
 type CommandPayload = {
   equation: string;
@@ -41,20 +38,22 @@ export default function EquationsPlugin(): JSX.Element | null {
   const [editor] = useLexicalComposerContext();
   const prevEquationEditorKeyRef = useRef<string | null>(null);
 
-  useEffect(() => {
-    hideActiveEquationEditor = () => {
-      editor.update(() => {
-        if (prevEquationEditorKeyRef.current) {
-          const node = $getNodeByKey(prevEquationEditorKeyRef.current);
-          if ($isEquationEditorNode(node)) {
-            node.hide();
-          }
-          prevEquationEditorKeyRef.current = null;
+  // Hide the active equation editor when the selection changes.
+  const hideActiveEquationEditor = () => {
+    editor.update(() => {
+      if (prevEquationEditorKeyRef.current) {
+        const node = $getNodeByKey(prevEquationEditorKeyRef.current);
+        if ($isEquationEditorNode(node)) {
+          node.hide();
         }
-      });
-    };
+        prevEquationEditorKeyRef.current = null;
+      }
+    });
+  };
 
-    return editor.registerUpdateListener(({editorState}) => {
+  // Listen for selection changes.
+  useEffect(() => {
+    return editor.registerUpdateListener(({ editorState }) => {
       editorState.read(() => {
         const selection = $getSelection();
         if ($isRangeSelection(selection)) {
@@ -104,19 +103,18 @@ export default function EquationsPlugin(): JSX.Element | null {
   return null;
 }
 
-function $getEquationEditorNodeOrNull(node: TextNode | ElementNode): EquationEditorNode | null {
+function $getEquationEditorNodeOrNull(
+  node: TextNode | ElementNode,
+): EquationEditorNode | null {
   if ($isEquationEditorNode(node)) {
-    console.log('equation editor node')
     return node;
   }
 
   if ($isTextNode(node) && $isEquationEditorNode(node.getParent())) {
-    console.log('text node')
     return node.getParent() as EquationEditorNode;
   }
 
   if ($isBlockEquationNode(node)) {
-    console.log('block equation node')
     return node.getFirstChild() as EquationEditorNode;
   }
